+++ csrc/aio/py_lib/deepspeed_aio_op_desc.cpp
@@ -11,8 +11,8 @@ io_op_desc_t::io_op_desc_t(const bool read_op,
                            const torch::Tensor& buffer,
                            const int fd,
                            const char* filename,
-                           const long long int file_num_bytes,
-                           const long long int file_offset,
+                           const int64_t file_num_bytes,
+                           const int64_t file_offset,
                            const int intra_op_parallelism,
                            const bool validate)
     : _read_op(read_op),
@@ -22,7 +22,7 @@ io_op_desc_t::io_op_desc_t(const bool read_op,
       _file_num_bytes(file_num_bytes),
       _file_offset(file_offset),
       _intra_op_parallelism(intra_op_parallelism),
-      _num_bytes_per_thread( static_cast<long long int>(buffer.nbytes())/ intra_op_parallelism),
+      _num_bytes_per_thread( static_cast<int64_t>(buffer.nbytes())/ intra_op_parallelism),
       _validate(validate)
 {
 }
