+++ csrc/aio/py_lib/deepspeed_aio_op_desc.h
@@ -14,10 +14,10 @@ struct io_op_desc_t {
     torch::Tensor _buffer;
     int _fd;
     const std::string _filename;
-    const long long int _file_num_bytes;
-    const long long int _file_offset;
+    const int64_t _file_num_bytes;
+    const int64_t _file_offset;
     const int _intra_op_parallelism;
-    const long long int _num_bytes_per_thread;
+    const int64_t _num_bytes_per_thread;
     torch::Tensor _contiguous_buffer;
     const bool _validate;
 
@@ -25,8 +25,8 @@ struct io_op_desc_t {
                  const torch::Tensor& buffer,
                  const int fd,
                  const char* filename,
-                 const long long int file_num_bytes,
-                 const long long int file_offset,
+                 const int64_t file_num_bytes,
+                 const int64_t file_offset,
                  const int intra_op_parallelism,
                  const bool validate);
