+++ csrc/aio/py_lib/deepspeed_cpu_op.h
@@ -18,8 +18,8 @@ struct cpu_op_desc_t : io_op_desc_t {
                   const std::unique_ptr<struct deepspeed_pin_tensor_t>& pinned_tensor_mgr,
                   const int fd,
                   const char* filename,
-                  const long long int file_num_bytes,
-                  const long long int file_offset,
+                  const int64_t file_num_bytes,
+                  const int64_t file_offset,
                   const int intra_op_parallelism,
                   const bool validate);
