+++ csrc/aio/py_lib/deepspeed_py_io_handle.h
@@ -45,7 +45,7 @@ struct deepspeed_io_handle_t {
     int pread(const torch::Tensor& buffer,
               const char* filename,
               const bool validate,
-              const long long int file_offset,
+              const int64_t file_offset,
               const bool async);
 
     int pwrite(const torch::Tensor& buffer,
@@ -53,11 +53,11 @@ struct deepspeed_io_handle_t {
                const bool validate,
                const bool async);
 
-    int sync_pread(torch::Tensor& buffer, const char* filename, const long long int file_offset);
+    int sync_pread(torch::Tensor& buffer, const char* filename, const int64_t file_offset);
 
     int sync_pwrite(const torch::Tensor& buffer, const char* filename);
 
-    int async_pread(torch::Tensor& buffer, const char* filename, const long long int file_offset);
+    int async_pread(torch::Tensor& buffer, const char* filename, const int64_t file_offset);
 
     int async_pwrite(const torch::Tensor& buffer, const char* filename);
 
@@ -82,7 +82,7 @@ struct deepspeed_io_handle_t {
         const torch::Tensor& buffer,
         const int fd,
         const char* filename,
-        const long long int file_num_bytes,
-        const long long int file_offset,
+        const int64_t file_num_bytes,
+        const int64_t file_offset,
         const bool validate);
 };
